+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
 
     <application
         android:allowBackup="true"
-        android:name=".app.App"
+        android:name=".App"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
