+++ app/src/main/java/com/movie/android/di/MainModule.kt
@@ -5,7 +5,7 @@ import com.movie.android.data.ApiInterceptor
 import com.movie.android.view.adapter.PopularMovieAdapter
 import com.movie.android.data.MainRepository
 import com.movie.android.data.ApiServices
-import com.movie.android.viewmodel.MainViewModel
+import com.movie.android.view.viewmodel.MainViewModel
 import okhttp3.OkHttpClient
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
