+++ app/src/main/java/com/movie/android/utils/MainUiState.kt
@@ -1,4 +1,4 @@
-package com.movie.android.viewmodel
+package com.movie.android.utils
 
 import com.movie.android.domain.Movie
