+++ app/src/main/java/com/movie/android/view/viewmodel/MainViewModel.kt
@@ -1,8 +1,9 @@
-package com.movie.android.viewmodel
+package com.movie.android.view.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.movie.android.data.MainRepository
+import com.movie.android.utils.MainUiState
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
